feat(plugins): Unable to setup github secrets which are needed for github actions to run after scaffolding (#138)
Unable to setup github secrets whicha re needed for github actions to
run after scaffolding

[ARC-116](https://github.com/sourcefuse/backstage/issues/116)

## Description
**After the microservices repo scaffolding, we need to add docker
related secrets to github for github actions to perform docker build and
docker push. This is not working because the repo which is created, the
creator from Backstage is not admin there by default.**

To resolve above issue, we should have provision to add collaborators
from templates. If collaborator not added from the template, provided
patch script. It would fetch current logged in github user's userId and
add it as collaborator with admin role.

![image](https://github.com/sourcefuse/backstage/assets/109595269/eff086b4-2179-494d-9bfb-2b765b1693b0)


https://github.com/sourcefuse/backstage/assets/109595269/bd7fe35b-d3cf-49d0-8b3d-280944f1011f

As per above video, I logged in with github id 'sadarunnisa-sf', which
is member of organization 'sadaru-tech-org1'. Once I created repo from
template, settings tab enabled and 'sadarunnisa-sf' was added as
collaborator with admin role
Fixes # (issue)
[ARC-116](https://github.com/sourcefuse/backstage/issues/116)

## Type of change

Please delete options that are not relevant.

- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Intermediate change (work in progress)

## How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration

- [ ] Test A
- [ ] Test B

## Checklist:

- [X] Performed a self-review of my own code
- [X] npm test passes on your machine
- [ ] New tests added or existing tests modified to cover all changes
- [ ] Code conforms with the style guide
- [ ] API Documentation in code was updated
- [ ] Any dependent changes have been merged and published in downstream
modules


[ARC-116]:
https://sourcefuse.atlassian.net/browse/ARC-116?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[ARC-116]:
https://sourcefuse.atlassian.net/browse/ARC-116?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ