feat(backend): Github based permission management (#206)
## Description

This PR introduces a new feature to the Backstage GitHub Permission
Plugin. It enables GitHub-based permissions to be mapped with
Backstage's permission framework. This allows for more granular control
and security when integrating GitHub permissions with Backstage
applications.

Fixes #187

## Type of change

Please delete options that are not relevant.

- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Intermediate change (work in progress)

## How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration

- [ ] Test A
- [ ] Test B

## Checklist:

- [X] Performed a self-review of my own code
- [ ] npm test passes on your machine
- [ ] New tests added or existing tests modified to cover all changes
- [ ] Code conforms with the style guide
- [ ] API Documentation in code was updated
- [X] Any dependent changes have been merged and published in downstream
modules